+++ addOns/authhelper/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Support for session management identification.
 - Support for auto-detect authentication.
 - Support for auto-detect session management.
+- Support for auto-detect verification.
 
 ### Fixed
 - Clear launched browser authentication when disabled, otherwise it would prevent enabling it again.